Change js files to tsx files
Modified all js files to tsx files
Added typescript suppressions to allow app to function. These
suppresions will begin being removed in the next commit
