+++ src/App.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import App from './App';
+import { App } from './App';
 
 export default {
     title: 'App Component',
